+++ CHANGELOG.md
@@ -1,9 +1,16 @@
 # Change Log
 
-## Version 0.0.3 (9 Dec 2024)
+## 0.0.4 (20 Feb 2025)
+
+- Add generic getters for single and array values
+- Add default values for getters
+- Separate access interfaces into generic and typed
+- Change flattened access control to method instead of assigning boolean field directly
+
+## 0.0.3 (9 Dec 2024)
 
 - Change package for public repository publishing
 
-## Version 0.0.2 (1 Oct 2024)
+## 0.0.2 (1 Oct 2024)
 
 - Update Config to be serializable